Dashboard Views and UI Changes

+++ src/main/java/com/ffg/rrn/model/Dashboard.java
@@ -41,105 +41,105 @@ public class Dashboard {
 	private Integer q4AssessmentComplete;
 
 	/* Referral REasons 1 to 14 */
-	private Integer rr1q1Count;
-	private Integer rr1q2Count;
-	private Integer rr1q3Count;
-	private Integer rr1q4Count;
+	private Long rr1q1Count;
+	private Long rr1q2Count;
+	private Long rr1q3Count;
+	private Long rr1q4Count;
 
-	private Integer rr2q1Count;
-	private Integer rr2q2Count;
-	private Integer rr2q3Count;
-	private Integer rr2q4Count;
+	private Long rr2q1Count;
+	private Long rr2q2Count;
+	private Long rr2q3Count;
+	private Long rr2q4Count;
 
-	private Integer rr3q1Count;
-	private Integer rr3q2Count;
-	private Integer rr3q3Count;
-	private Integer rr3q4Count;
+	private Long rr3q1Count;
+	private Long rr3q2Count;
+	private Long rr3q3Count;
+	private Long rr3q4Count;
 
-	private Integer rr4q1Count;
-	private Integer rr4q2Count;
-	private Integer rr4q3Count;
-	private Integer rr4q4Count;
+	private Long rr4q1Count;
+	private Long rr4q2Count;
+	private Long rr4q3Count;
+	private Long rr4q4Count;
 
-	private Integer rr5q1Count;
-	private Integer rr5q2Count;
-	private Integer rr5q3Count;
-	private Integer rr5q4Count;
+	private Long rr5q1Count;
+	private Long rr5q2Count;
+	private Long rr5q3Count;
+	private Long rr5q4Count;
 
-	private Integer rr6q1Count;
-	private Integer rr6q2Count;
-	private Integer rr6q3Count;
-	private Integer rr6q4Count;
+	private Long rr6q1Count;
+	private Long rr6q2Count;
+	private Long rr6q3Count;
+	private Long rr6q4Count;
 
-	private Integer rr7q1Count;
-	private Integer rr7q2Count;
-	private Integer rr7q3Count;
-	private Integer rr7q4Count;
+	private Long rr7q1Count;
+	private Long rr7q2Count;
+	private Long rr7q3Count;
+	private Long rr7q4Count;
 
-	private Integer rr8q1Count;
-	private Integer rr8q2Count;
-	private Integer rr8q3Count;
-	private Integer rr8q4Count;
+	private Long rr8q1Count;
+	private Long rr8q2Count;
+	private Long rr8q3Count;
+	private Long rr8q4Count;
 
-	private Integer rr9q1Count;
-	private Integer rr9q2Count;
-	private Integer rr9q3Count;
-	private Integer rr9q4Count;
+	private Long rr9q1Count;
+	private Long rr9q2Count;
+	private Long rr9q3Count;
+	private Long rr9q4Count;
 
-	private Integer rr10q1Count;
-	private Integer rr10q2Count;
-	private Integer rr10q3Count;
-	private Integer rr10q4Count;
+	private Long rr10q1Count;
+	private Long rr10q2Count;
+	private Long rr10q3Count;
+	private Long rr10q4Count;
 
-	private Integer rr11q1Count;
-	private Integer rr11q2Count;
-	private Integer rr11q3Count;
-	private Integer rr11q4Count;
+	private Long rr11q1Count;
+	private Long rr11q2Count;
+	private Long rr11q3Count;
+	private Long rr11q4Count;
 
-	private Integer rr12q1Count;
-	private Integer rr12q2Count;
-	private Integer rr12q3Count;
-	private Integer rr12q4Count;
+	private Long rr12q1Count;
+	private Long rr12q2Count;
+	private Long rr12q3Count;
+	private Long rr12q4Count;
 
-	private Integer rr13q1Count;
-	private Integer rr13q2Count;
-	private Integer rr13q3Count;
-	private Integer rr13q4Count;
+	private Long rr13q1Count;
+	private Long rr13q2Count;
+	private Long rr13q3Count;
+	private Long rr13q4Count;
 
-	private Integer rr14q1Count;
-	private Integer rr14q2Count;
-	private Integer rr14q3Count;
-	private Integer rr14q4Count;
+	private Long rr14q1Count;
+	private Long rr14q2Count;
+	private Long rr14q3Count;
+	private Long rr14q4Count;
 
-	private Integer housingQ1Count;
-	private Integer housingQ2Count;
-	private Integer housingQ3Count;
-	private Integer housingQ4Count;
+	private Long housingQ1Count;
+	private Long housingQ2Count;
+	private Long housingQ3Count;
+	private Long housingQ4Count;
 
-	private Integer mmQ1Count;
-	private Integer mmQ2Count;
-	private Integer mmQ3Count;
-	private Integer mmQ4Count;
+	private Long mmQ1Count;
+	private Long mmQ2Count;
+	private Long mmQ3Count;
+	private Long mmQ4Count;
 
-	private Integer empQ1Count;
-	private Integer empQ2Count;
-	private Integer empQ3Count;
-	private Integer empQ4Count;
+	private Long empQ1Count;
+	private Long empQ2Count;
+	private Long empQ3Count;
+	private Long empQ4Count;
 
-	private Integer eduQ1Count;
-	private Integer eduQ2Count;
-	private Integer eduQ3Count;
-	private Integer eduQ4Count;
+	private Long eduQ1Count;
+	private Long eduQ2Count;
+	private Long eduQ3Count;
+	private Long eduQ4Count;
 
-	private Integer nsQ1Count;
-	private Integer nsQ2Count;
-	private Integer nsQ3Count;
-	private Integer nsQ4Count;
+	private Long nsQ1Count;
+	private Long nsQ2Count;
+	private Long nsQ3Count;
+	private Long nsQ4Count;
 
-	private Integer hhQ1Count;
-	private Integer hhQ2Count;
-	private Integer hhQ3Count;
-	private Integer hhQ4Count;
+	private Long hhQ1Count;
+	private Long hhQ2Count;
+	private Long hhQ3Count;
+	private Long hhQ4Count;
 
 	private Long residentServedQ1;
 	private Long residentServedQ2;
